fix interception (mostly)
The forward declarations in asan_interceptors.h are probably not
supposed to be there. also interception_cheerpwasm.* can probably be
removed.

The code needs to still be refactored. Also probably a good idea to
setup CI for this as it is starting to get big, and it would be good to
test it on other systems.

This code requires https://github.com/Hyxogen/cheerp-musl/commit/5a7a1296c877f051b7b4b07588eaf3e18aa7f899
to work.