remove cmake hacks
It seems to be able to compiler libasan with multiple threads, however
I'm not convinced that the issue of corrupted bitcode is solved, when
trying to use ninja it complains that there are multiple rules for
generating a file, which it doesn't do on native.

My guess is that it is a race condition, but now that it also compilers
the dynamic library, the race doesn't occur.

As for llvm-nm, installing it allows less modifications to the
compiler-rt cmakelists and also honestly seems to me something usefull
to install.