gda: simplify invoke calls
The `simplifyCalls` routine in the GDA does not only "simplify calls"
but also replaces calls of bitcasts with bitcasts of the call. While
working on supporting exceptions with ASan, there was some generated
code where an invoke happened on a bitcast of a function. Because the
bitcast did not happen on the invoke, it later resulted in the (wrong)
assumption that this call was an indirect call of an unsupported type.
Causing the call to be replaced by an unreachable instruction.

Made it so that the `simplifyCalls` routine runs on CallBase.